move setPlayers into separate file and improve it
